chore: update structure by moving separate components to common components folder
